Excersise 8 fixed

+++ csharp-basics/exercises/TypesAndVariables/Excersise 8/Program.cs
@@ -7,10 +7,10 @@ namespace Excersise_8
         static void Main(string[] args)
         {
             Console.WriteLine("Please enter a number of minutes");
-            int minutes = int.Parse(Console.ReadLine());
+            double minutes = double.Parse(Console.ReadLine());
 
-            int years = minutes / 525600;
-            int days = minutes / 1440;
+            double years = minutes / 525600;
+            double days = minutes / 1440;
 
             Console.WriteLine("{0} minutes are {1} years or {2} days", minutes, years, days);
             Console.ReadKey();
